+++ app/src/main/java/cmc/goalmate/app/navigation/NavigationTypeAliases.kt
@@ -1,3 +1,3 @@
 package cmc.goalmate.app.navigation
 
-typealias NavigateToGoal = (goalId: Long) -> Unit
+typealias NavigateToGoal = (goalId: Int) -> Unit
