+++ app/src/main/java/cmc/goalmate/presentation/ui/progress/inprogress/InProgressAction.kt
@@ -14,7 +14,7 @@ sealed interface InProgressAction {
 
     data object ClickUneditableGoal : InProgressAction
 
-    data class NavigateToComments(val goalId: Long) : InProgressAction
+    data class NavigateToComments(val goalId: Int) : InProgressAction
 
-    data class NavigateToGoalDetail(val goalId: Long) : InProgressAction
+    data class NavigateToGoalDetail(val goalId: Int) : InProgressAction
 }
